Implement audio effects and fix critical game mechanics in poker game
This commit introduces sound effects and resolves major issues with the game mechanics, particularly in heads-up play and all-in situations. The changes aim to create a more immersive and intuitive gaming environment while ensuring the game's mechanics function properly.

Audio Effects:
- Added sound effects to the moving elements of the game, such as cards, chips, and the dealer button, enriching the overall user experience.
- Implemented a sound effect for when a player is eliminated, adding to the game's audio feedback.

Major Bug Fixes in Game Mechanics of Headsup Play:
- Fixed a critical issue where a player going all-in to post a blind was not being dealt hole cards. This change ensures that all players with stack or total bet receive hole cards.
- Resolved a problem where the betting round would end prematurely if one player went all-in to post a blind. Now, the round only ends after the other player has responded to the current bet, thanks to the refactored 'is_round_over()' conditional.
- Addressed a similar issue in heads-up play where the round did not end after one player went all-in to post a blind and the other went all-in. This was achieved by refactoring 'is_single_active_player()' to 'is_less_two_active_players()', ensuring the game correctly identifies scenarios with no active players.

Minor Game Mechanics Enhancements and Fixes:
- Prevented creation of unneeded side pots (related to the last issue). Previously, players had the option to raise or go all-in even when all other players were not active (i.e., could not respond to the raise). These actions are no longer available when there are no other active players.
- Streamlined player rotation. The BB turn is now skipped to end the pre-flop betting round if he has the option and there are no other active players.
- Fixed an issue where not all the hand results were displayed.
- Corrected the display of the action label when a player goes all-in to post a blind.

GUI Improvements:
- Added a dollar sign ('$') to the pot label text for better clarity in betting.
- Refactored 'main.py' by composing a class from the Pyglet window and related functions for improved code organization and readability.

These changes significantly improve the game's reliability and playability.

- Thoroughly tested new audio features and game mechanics fixes.

+++ src/gui/player.py
@@ -1,3 +1,5 @@
+import random
+
 import pyglet
 
 from src.core.constants import Action
@@ -5,6 +7,8 @@ from src.gui import load
 from random import randint
 from src.gui.card import CardSprite
 from src.gui.constants import PLAYER_POSITIONS, OFF_SCREEN
+from src.gui.resources import card_shoving_sounds, chip_stacking_sounds, chip_shoving_sound, tapping_sounds, \
+    elimination_sound
 
 
 class PlayerSprite:
@@ -12,6 +16,7 @@ class PlayerSprite:
         self.player = player
         self.x, self.y = x, y
         self.batch = batch
+        self.stack_size = player.stack_size
 
         self.name_x, self.name_y = x + 70, y + 90
         self.chip_x, self.chip_y = x - 80, y + 50
@@ -56,21 +61,42 @@ class PlayerSprite:
     def update(self):
         # name
         if not self.player.total_bet and not self.player.stack_size and self.player.action != Action.ALL_IN:
-            self.name_label.x = OFF_SCREEN
+            if self.name_label.x != OFF_SCREEN:
+                # play elimination sound
+                elimination_sound.play()
+
+                self.name_label.x = OFF_SCREEN
         else:
             self.name_label.x = self.name_x
 
         # stack
-        if self.player.stack_size:
-            self.stack_size_label.text = f'${str(self.player.stack_size)}'
-            self.chip_sprite.x = self.chip_x
-        else:
-            self.stack_size_label.text = ''
-            self.chip_sprite.x = OFF_SCREEN
+        if self.stack_size != self.player.stack_size:
+            # play chip stack sound
+            if self.stack_size < self.player.stack_size:
+                random.choice(chip_stacking_sounds).play()
+
+            self.stack_size = self.player.stack_size
+
+            if self.player.stack_size:
+                self.stack_size_label.text = f'${str(self.stack_size)}'
+                self.chip_sprite.x = self.chip_x
+            else:
+                self.stack_size_label.text = ''
+                self.chip_sprite.x = OFF_SCREEN
 
         # action
         if self.player.action not in [None, Action.POST_BLIND] and (self.player.stack_size or self.player.hole_cards):
-            self.action_label.text = self.player.action.value
+            if self.action_label.text != self.player.action.value:
+                # play action sound (except folding)
+                if self.player.action in [Action.BET, Action.RAISE, Action.CALL]:
+                    random.choice(chip_stacking_sounds).play()
+                elif self.player.action == Action.CHECK:
+                    random.choice(tapping_sounds).play()
+                    pyglet.clock.schedule_once(lambda dt: random.choice(tapping_sounds).play(), 1 / 5)
+                elif self.player.action == Action.ALL_IN:
+                    chip_shoving_sound.play()
+
+                self.action_label.text = self.player.action.value
         else:
             self.action_label.text = ''
 
@@ -78,6 +104,9 @@ class PlayerSprite:
         if self.player.hole_cards:
             for card_int in self.player.hole_cards:
                 if card_int not in self.card_sprites:
+                    # play card shoving sound
+                    random.choice(card_shoving_sounds).play()
+
                     self.card_sprites[card_int] = CardSprite(card_int,
                                                              x=self.card_x + self.player.hole_cards.index(
                                                                  card_int) * 25,
@@ -90,7 +119,10 @@ class PlayerSprite:
                     self.card_sprites[card_int].face_up()
                 else:
                     self.card_sprites[card_int].face_down()
-        else:
+        elif self.card_sprites:
+            # play card shoving sound
+            random.choice(card_shoving_sounds).play()
+
             for k, sprite in self.card_sprites.items():
                 sprite.delete()
             self.card_sprites.clear()
